POSHI-107 Move pql from 'poshi-runner' into 'poshi-core'
