+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/pql/PQLEntity.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.poshi.runner.pql;
+package com.liferay.poshi.core.pql;
 
 import java.util.Properties;
 import java.util.Set;
