+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/pql/PQLModifierFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.poshi.runner.pql;
+package com.liferay.poshi.core.pql;
 
 /**
  * @author Michael Hashimoto
