+++ modules/test/poshi/poshi-core/src/test/java/com/liferay/poshi/core/pql/PQLOperatorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.poshi.runner.pql;
+package com.liferay.poshi.core.pql;
 
 import java.util.HashSet;
 import java.util.Properties;
