+++ modules/test/poshi/poshi-core/src/test/java/com/liferay/poshi/core/pql/PQLValueTest.java
@@ -12,9 +12,11 @@
  * details.
  */
 
-package com.liferay.poshi.runner.pql;
+package com.liferay.poshi.core.pql;
 
+import java.util.HashSet;
 import java.util.Properties;
+import java.util.Set;
 
 import junit.framework.TestCase;
 
@@ -23,7 +25,7 @@ import org.junit.Test;
 /**
  * @author Michael Hashimoto
  */
-public class PQLVariableTest extends TestCase {
+public class PQLValueTest extends TestCase {
 
 	@Test
 	public void testGetPQLResult() throws Exception {
@@ -55,41 +57,67 @@ public class PQLVariableTest extends TestCase {
 
 	@Test
 	public void testGetPQLResultError() throws Exception {
-		_validateGetPQLResultError(
-			"invalid.property", "Invalid testcase property: invalid.property");
-		_validateGetPQLResultError(null, "Invalid variable: null");
-		_validateGetPQLResultError(
-			"test == test", "Invalid value: test == test");
-		_validateGetPQLResultError(
-			"test OR test", "Invalid value: test OR test");
+		Set<String> pqls = new HashSet<>();
+
+		pqls.add("test == test");
+		pqls.add("test test");
+		pqls.add("true AND true");
+
+		for (String pql : pqls) {
+			_validateGetPQLResultError(pql, "Invalid value: " + pql);
+		}
 	}
 
-	private void _validateGetPQLResult(String pql, Object expectedPQLResult)
+	@Test
+	public void testGetPQLResultModifier() throws Exception {
+		_validateGetPQLResult("NOT true", Boolean.FALSE);
+		_validateGetPQLResult("NOT false", Boolean.TRUE);
+	}
+
+	@Test
+	public void testGetPQLResultModifierError() throws Exception {
+		_validateGetPQLResultError(
+			"NOT 3.2", "Modifier must be used with a boolean value: NOT");
+		_validateGetPQLResultError(
+			"NOT 2016", "Modifier must be used with a boolean value: NOT");
+		_validateGetPQLResultError(
+			"NOT test", "Modifier must be used with a boolean value: NOT");
+		_validateGetPQLResultError(
+			"NOT 'test test'",
+			"Modifier must be used with a boolean value: NOT");
+	}
+
+	@Test
+	public void testGetPQLResultNull() throws Exception {
+		_validateGetPQLResultNull(null);
+		_validateGetPQLResultNull("'null'");
+		_validateGetPQLResultNull("\"null\"");
+	}
+
+	private void _validateGetPQLResult(String pql, Object expectedResult)
 		throws Exception {
 
 		Properties properties = new Properties();
 
-		properties.put("portal.smoke", pql);
+		Class<?> clazz = expectedResult.getClass();
 
-		Class<?> clazz = expectedPQLResult.getClass();
+		PQLValue pqlValue = new PQLValue(pql);
 
-		PQLVariable pqlVariable = new PQLVariable("portal.smoke");
+		Object actualResult = pqlValue.getPQLResult(properties);
 
-		Object actualPQLResult = pqlVariable.getPQLResult(properties);
-
-		if (!clazz.isInstance(actualPQLResult)) {
+		if (!clazz.isInstance(actualResult)) {
 			throw new Exception(pql + " should be of type: " + clazz.getName());
 		}
 
-		if (!actualPQLResult.equals(expectedPQLResult)) {
+		if (!actualResult.equals(expectedResult)) {
 			StringBuilder sb = new StringBuilder();
 
 			sb.append("Mismatched PQL result within the following PQL:\n");
 			sb.append(pql);
 			sb.append("\n* Actual:   ");
-			sb.append(actualPQLResult);
+			sb.append(actualResult);
 			sb.append("\n* Expected: ");
-			sb.append(expectedPQLResult);
+			sb.append(expectedResult);
 
 			throw new Exception(sb.toString());
 		}
@@ -101,9 +129,9 @@ public class PQLVariableTest extends TestCase {
 		String actualError = null;
 
 		try {
-			PQLVariable pqlVariable = new PQLVariable(pql);
+			PQLValue pqlValue = new PQLValue(pql);
 
-			pqlVariable.getPQLResult(new Properties());
+			pqlValue.getPQLResult(new Properties());
 		}
 		catch (Exception exception) {
 			actualError = exception.getMessage();
@@ -111,7 +139,7 @@ public class PQLVariableTest extends TestCase {
 			if (!actualError.equals(expectedError)) {
 				StringBuilder sb = new StringBuilder();
 
-				sb.append("Mismatched error within the following PQL:\n");
+				sb.append("Mismatched error for the following PQL:\n");
 				sb.append(pql);
 				sb.append("\n* Actual:   ");
 				sb.append(actualError);
@@ -126,6 +154,29 @@ public class PQLVariableTest extends TestCase {
 				throw new Exception(
 					"No error thrown for the following PQL: " + pql);
 			}
+		}
+	}
+
+	private void _validateGetPQLResultNull(String pql) throws Exception {
+		Properties properties = new Properties();
+
+		PQLValue pqlValue = new PQLValue(pql);
+
+		Object actualResult = pqlValue.getPQLResult(properties);
+
+		if (actualResult != null) {
+			Object expectedResult = null;
+
+			StringBuilder sb = new StringBuilder();
+
+			sb.append("Mismatched PQL result within the following PQL:\n");
+			sb.append(pql);
+			sb.append("\n* Actual:   ");
+			sb.append(actualResult);
+			sb.append("\n* Expected: ");
+			sb.append(expectedResult);
+
+			throw new Exception(sb.toString());
 		}
 	}
