+++ modules/test/poshi/source-formatter-suppressions.xml
@@ -2,6 +2,7 @@
 
 <suppressions>
 	<checkstyle>
+		<suppress checks="BuilderCheck" files="poshi-core/.*" />
 		<suppress checks="BuilderCheck" files="poshi-runner/.*" />
 	</checkstyle>
 	<source-check>
@@ -10,8 +11,8 @@
 		<suppress checks="JavaModuleTestCheck" files="poshi-runner/src/main/java/com/liferay/poshi/runner/PoshiRunner\.java" />
 		<suppress checks="JavaModuleTestCheck" files="poshi-runner/src/main/java/com/liferay/poshi/runner/PoshiRunnerTestCase\.java" />
 		<suppress checks="JavaMultiPlusConcatCheck" files=".*" />
-		<suppress checks="JavaStylingCheck" files="poshi-runner/src/test/java/com/liferay/poshi/runner/pql/PQLEntityFactoryTest\.java" />
-		<suppress checks="JavaStylingCheck" files="poshi-runner/src/test/java/com/liferay/poshi/runner/pql/PQLQueryTest\.java" />
+		<suppress checks="JavaStylingCheck" files="poshi-core/src/test/java/com/liferay/poshi/core/pql/PQLEntityFactoryTest\.java" />
+		<suppress checks="JavaStylingCheck" files="poshi-core/src/test/java/com/liferay/poshi/core/pql/PQLQueryTest\.java" />
 		<suppress checks="JavaXMLSecurityCheck" files=".*" />
 		<suppress checks="LocaleUtilCheck" files="poshi-runner/src/main/java/com/liferay/poshi/runner/util/LocaleUtil\.java" />
 		<suppress checks="MarkdownWhitespaceCheck" files="poshi-script\.markdown" />
